Connector: removed socket.shutdown()
The .shutdown() call can only be performed when using a connection oriented transport protocol (here: TCP/SOCK_STREAM).
It sends a message with the FIN flag to terminate the connection from our end. After that, we can still (only) receive data from the socket and it is up to the other end to end the connection with FIN. (though, here we used SHUT_RDWR so no reading would be possible)
Since we close the socket right afterwards, we don't need to call .shutdown().
Calling .shutdown() could also potentially cause issues if the socket was shared with another process.